Create blendshape_tgt_map_importer.py
first attempt and mel.eval importer for blendshapes. 
You can select a mesh with blendshape targets on it, and run this. It allows you to import maps in any image format(preferably .tga), and split the targets based on the maps. 
For this to work, the prefix of the map file has to match the prefix of the target.